Moved CurrentAndNextEvents NoEvents to their own file
