+++ client/app/components/conference_room/ConferenceRoomContainer.jsx
@@ -1,47 +1,14 @@
-import moment from 'moment';
-import get from 'lodash/get';
 import React from 'react';
 import EventSchema from 'proptypes/schemas/EventSchema';
 import ConferenceRoomSchema from 'proptypes/schemas/ConferenceRoomSchema';
 import { Grid, Row, Col, Jumbotron } from 'react-bootstrap';
 
 import Navbar from './layout/Navbar';
-import CurrentEvent from './event/CurrentEvent';
-import NextEvents from './event/NextEvents';
+import CurrentAndNextEvents from './layout/CurrentAndNextEvents'
+import NoEvents from './layout/NoEvents'
+
 
 import './conference_room.scss';
-
-const NO_MORE_EVENTS_TEXT = 'No more events for today';
-
-const CurrentAndNextEvents = ({ currentEvent, nextEvents, onUpdate }) => (
-  <Row>
-    <Col xs={12} sm={8}>
-      <CurrentEvent event={currentEvent}
-                    nextEventStart={moment(get(nextEvents[0], 'start.date_time'))}
-                    onCompleted={onUpdate} />
-    </Col>
-    <Col xs={12} sm={4}>
-      <NextEvents events={nextEvents} noEventLabel={NO_MORE_EVENTS_TEXT} />
-    </Col>
-  </Row>
-);
-
-CurrentAndNextEvents.propTypes = {
-  currentEvent: EventSchema.except('width', 'offset'),
-  nextEvents: React.PropTypes.arrayOf(EventSchema.only('start')),
-  onUpdate: React.PropTypes.func
-};
-
-const NoEvents = () => (
-  <Row>
-    <Col xs={12}>
-      <Jumbotron>
-        <h3>{NO_MORE_EVENTS_TEXT}</h3>
-        <p>Check back tomorrow!</p>
-      </Jumbotron>
-    </Col>
-  </Row>
-);
 
 const ConferenceRoomContainer = ({
   currentEvent,
