+++ client/app/components/conference_room/texts.js
@@ -1,0 +1,3 @@
+export default {
+  NO_MORE_EVENTS: 'No more events for today'
+};
